NetworkParameters: change `packetMagic` to an `int`
It's just a binary value and does not have signedness.

Methods for writing 32-bit integers via `ByteUtils.writeInt32BE(int, ...)` have been added.

+++ core/src/main/java/org/bitcoinj/core/NetworkParameters.java
@@ -69,7 +69,7 @@ public abstract class NetworkParameters {
 
     protected BigInteger maxTarget;
     protected int port;
-    protected long packetMagic;  // Indicates message origin network and is used to seek to the next message when stream state is unknown.
+    protected int packetMagic;  // Indicates message origin network and is used to seek to the next message when stream state is unknown.
     protected int addressHeader;
     protected int p2shHeader;
     protected int dumpedPrivateKeyHeader;
@@ -300,7 +300,7 @@ public abstract class NetworkParameters {
      * The header bytes that identify the start of a packet on this network.
      * @return header bytes as a long
      */
-    public long getPacketMagic() {
+    public int getPacketMagic() {
         return packetMagic;
     }
